Add a fallback for simulator server image when no is provided (#1309)
This PR adds a fallback for when simulator server is killed but the
webview still tries to read the stream.

To do this in a way that will not disrupt our ability to diagnose
problems within the IDE we introduce new `IS_DEV` value passed down to
the webview and conditionally render the visible (and annoying) "NO
IMAGE" warning if we are in the development mode.

In production we replace the stream with a black screen to mask any
imperfections in our state management and provide better user
experience.

Dev: 

<img width="347" alt="Screenshot 2025-07-08 at 12 22 42"
src="https://github.com/user-attachments/assets/cede927a-57ab-4104-98db-7dbd9d0ffb2b"
/>

Production: 


https://github.com/user-attachments/assets/e81cf7bd-a84e-4fc8-8820-7cf70e6da321

### How Has This Been Tested: 

- Run any test application and wait for the stream to be visible 
- find `simulator-server-macos` process in activity monitor and kill it 
- do it both for prod and dev environment.

### How Has This Change Been Documented:

Not applicable