Delete theme type setting (#1324)
This PR deletes all the code that was supporting the theme type
selection.

Historically, we had our own theme for the panel UI for light and dark
editor themes. A while ago, we added a way to match the editor theme
which in most of the cases look much better and consistent. Now, with us
adapting vscode-elements component library, we don't want to fight the
vscode matching elements any longer such that they look ok on the theme
not many people use.

This PR phases out the support for the setting while only keeping it in
in the setting scheme with the deprecation notice in case anyone still
has it set.

There was one variable `--swm-select-item-disabled-background` that was
present in the built-in themes but missing in the vscode ones. I
investigated this variable and found out it was used by the select
dropdown for disabled options. I consulted the existing vscode
components, which instead of changing the background for disabled items,
change the foreground only. Finally, I ended up updating the stylesheet
to no longer use that variable, and instead use a proper disabled
foreground color for disable select items.

### How Has This Been Tested: 
1) Open Radon IDE and see it look ok with the editor theme
2) Set the theme setting to "built-in" and make sure it falls back to
the editor theme.

---------

Co-authored-by: Juliusz Wajgelt <49338439+jwajgelt@users.noreply.github.com>