Fix Inspector while using the caught exceptions functionality  (#1335)
This PR fixes an issue caused by internal inspector implementation that
in order to get components details throws an error as this component.
This unfortunately interfered with the debugger breaking on those
exceptions. To fix this we skip aby breaks that ar caused by something
that was called by `getInspectorDataForViewAtPoint`.

Fixes #1333 

### How Has This Been Tested: 

- run test app on android 
- enable caught exceptions
- try and us an inspector