add launch env to build cache key (#1336)
Adds the `env` set up in the launch configuration to the build cache key
to ensure the app rebuilds when the `env` setting changes.
This is necessary for setups where the user wants to switch between two
different configs for the same application, e.g. an env variable which
enables/disables Storybook.

### How Has This Been Tested: 
- open `expo-router` test app
- add following launch configs:
```
{
      "name": "ruter",
      "type": "radon-ide",
      "request": "launch",
      "appRoot": "./expo-router"
    },
    {
      "name": "ruter (storybook)",
      "type": "radon-ide",
      "request": "launch",
      "appRoot": "./expo-router",
      "env": {
        "STORYBOOK_ENABLED": "true"
      }
    }
```
- check that the configs launch the app with/without storybook