feat: load environment variables from project`s dotenv files (#1361)
Loads the environment variables for the application root using
`@expo/env` and appends it to the system and configured environment that
is passed to the other parts of the extension.
Also changes the build cache key to use a hash of the environment,
instead of the full stringified env.

Closes #1350 

### How Has This Been Tested: 
- check that adding a `.env` file in the app root folder results in the
variables being added to the `ApplicationContext.launchConfig` object
- check that `.env` contents don't override values set in the system or
custom launch config
- check that switching app roots doesn't carry over the variables loaded
from the previously chosen approot