fix: saving initial launch configuration when it's detected (#1365)
This PR fixes an issue with detected launch configuration never being
used as initial launch configuration.
This was caused by the initial config detector always checking if the
saved configuration is still present in the custom configurations list
even if the configuration wan never part of that list.

To fix that we first check if the configurations kind is `Detected` and
if it is and the selected appRoot still exists we select it.

### How Has This Been Tested: 

- run `radon-ide-test-apps` directory 
- select one of the detected configs 
- turn of vscode 
- turn it back on and check if the above config is the new initial
config

### How Has This Change Been Documented:

Internal fix