refactor: move dependency management to the new state system  (#1375)
This PR moves the old DependencyManager and DependenciesProvider
functionality to the new state solution and introduces the following
changes as the result:

1) we devide `DependencyManager` into `EnvironmentDependencyManager` and
`ApplicationDependencyManager` to isolate dependencies that require
application specific information in the application context
2) we move the functionality of the package manager into a separate
callas to further improve the overblown `ApplicationDependencyManager`
part of `DependencyManager` while doing so we introduce similar patterns
as `pods` class with dedicated cancelToken and output channel.
3) we introduce `useDependencyErrors` to provide a functionality
previously defined in DependencyProvider
4) we move the `runAllChecks` method into project which shapes to be our
stable API for the backend part of the extension and remove
`DependenciesProvider`


5) I fix my mistake in `mergeAndCalculateChanges` from the previous
refactor when I assumed that `oldNode` always contains the necessary
key.

### How Has This Been Tested: 

- run `react-native-81` test app 
- remove node_modules and see them installed
- brak something like for example remove the android dir and check if it
shows up in the doctor

### How Has This Change Been Documented:

This is internal change