Initial setup for UI end to end testing (#1377)
This pull request introduces initial setup for end-to-end (E2E) tests
for the VS Code extension using the `vscode-extension-tester` framework.
The purpose of this change is to improve test coverage and ensure the
core functionality of the extension is validated in a real-world
environment.

Fixes # (issue)

### How Has This Been Tested: 

The added UI tests have been run on local development environment and
manually checked to ensure they work correctly. Here's the outline of
the testing process:

- **Local Testing**: Ran via script `npm run setup-run-tests` test cases
on local machine using the VS Code Extension Tester - all executed tests
passed successfully.

### How Has This Change Been Documented:

Currently, there is no documentation available. Documentation will be
provided in the future as more tests are added.

### Setup

Installation:

- NodeJS [v22.16.0]
- Visual Studio Code [1.101.2]
- ChromeDriver [138.0.7204.49]
- Python [3.13.5]
- selenium-webdriver [4.34.0]
- mocha [11.7.1]
- chai [5.2.0]
- vscode-extension-tester
- Prepare and mock test data:

Inside the data folder add a vsix radon file (with the version you want
to test) named `react-native-ide-1.9.0.vsix`. If you want to name it
differently, rename the vsix file name in `package.json` in the script
`install-vsix`
Inside the 'data' folder add a new folder named radon-ide-test-apps
which will contain the test applications folders. In my case,
`react-native-80` is there. If the folder with apps or the name of the
folder with the tested app itself has a different name - change it in
testData.js inside te data folder. This file also contains the expected
name of the project (in my case reactNative80) - change it if your
application is named differently.
Run setup and tests:

After preparing the data, run the command `npm run setup-run-tests`
which performs the following actions: retrieves Visual Studio Code and
ChromeDriver, installs the `./data/radon-ide.vsix` and runs each tests
inside the `./ui-tests/*.test.js`