refactor: remove utils interface (#1405)
This PR continues the work on uniform extension API by moving all the
`UtilisInterface` methods into `ProjectInterface`

to achieve this cleanly we introduce a new `EditorManager` class that
contains the logic implementing the large part of the old
`UtilsInterface`.

note: this PR is dependent on #1404

to do: 
- [ ] move the logic implementing the recording/replay/screenshot
functionality to a separate class.

### How Has This Been Tested: 

-run `react-native-81` test app and play with it 

### How Has This Change Been Documented:

Internal change