fix: node version check (#1414)
This PR changes `getMinimumSupportedNodeVersion` logic to return the
most restrictive version instead of the first found one. This Solves an
issue with projects that have updated rn version but not the engine
requirements in their package.json.

### How Has This Been Tested: 

add a brak point to the `getMinimumSupportedNodeVersion` method run in
tin the `react-native-81` test app and check if the return value is
22.14 instead of 18

### How Has This Change Been Documented:

internal change