feat: show alert when DevTools disconnect with information that tools may not work (#1413)
When the Android application is moved to the background, the Websockets
it opened all close. This causes the React DevTools connection to close,
breaking Radon's Inspector Bridge which is needed for many tools to
work.
This PR shows an alert when that happens to inform the user that
features may not work correctly and prompt them to restart the app.
Related #1400

<img width="528" height="478" alt="image"
src="https://github.com/user-attachments/assets/6beb71bf-673b-430f-a21d-6686c31350a9"
/>


### How Has This Been Tested: 
- open an app in Android and go to the launcher / switch application --
the alert should appear after a few seconds
- close the app fully on Android/iOS -- the alert should appear after
the app is killed
- open the app again, through the phone's launcher -- the alert should
disappear and tools should work again
- trigger any restart action while the alert is shown -- the alert
should disappear