fix: changing panel location  (#1442)
This PR solves an issue introduces by
https://github.com/software-mansion/radon-ide/pull/1147.

when the user changes the location of the IDE from tab to side panel
using `user` scope instead of the default `workspace`
they could find themselves in a illegal state of having radon open in a
tab panel when the setting would be set to
`side-panel`. Then if the user used an option to open IDE panel like
`preview` or just `show IDE` command. they would have to panels open at
once.

### How Has This Been Tested: 

-run a test app and open radon in tab panel 
- change it to a different app and switch "panel location" setting to
side-panel with `user` scope
- return to the first app and it will open in tab panel instead of side
panel

### How Has This Change Been Documented:

internal