feat: allow sending files to device (#1453)
Allows sending files to the selected device via Radon.
- adds a "Send file" option to the device settings dropdown which opens
the system file picker to select the file(s) to send
- allows drag-n-dropping files over the preview in order to send them
over to the device

UI to be done in a separate PR.


https://github.com/user-attachments/assets/afd426d1-53a0-471d-8dd6-4cb6ab65ddf2

Limitations: 
- currently, only media files are supported on iOS due to issues with
the "Files" app not working correctly in Radon Simulators. The default
behaviour of the Apple Simulator app when dropping a (non-media) file
over the simulator seems to be the same as `simctl openurl
file://${path-to-file}` which opens the "Files" app on the device and
allows selecting where the file should be saved. However, on the Radon
Simulator, the "Files" app is unable to save the file (the button is
disabled), and other functions (e.g. creating a new folder) also don't
work -- I don't know why.

### How Has This Been Tested: 
- open the `react-native-image-picker` example app
- use "send files" option in the Device Settings dropdown and select a
photo
- drag another photo to the preview
- verify both photos are selectable in the image picker

### How Has This Change Been Documented:
WIP