feat: add fps recording feature  (#1459)
NOTE: 
After internal changes we decided not to go with this approach but we
still wan to merge it as it contains useful info-structure that we can
use in the future. This PR does not expose any UI option to use the
frame rate repoerter.

This PR adds a new fps reporting functionality enabled by
https://github.com/software-mansion-labs/simulator-server/commit/f22a46abeafee0055687f39ea0fc2f16399c3824.
To do this we:

- integrate with the new sim server function 
- update our new state management system to allow for having states of
multiple device sessions at once (note: the `use$` function form legend
state seems to only work correctly when observing primitives if the
object in question is not part of the initial observable for example
then we use `Record` type. This is not a big limitation but will require
the structure of the frontend to look close to `useFrameReporting` hook)
- update the UI to allow for frame rate inspecting. 


https://github.com/user-attachments/assets/b69e2c06-303f-4ef1-b775-31f6e96dbdc6

### How Has This Been Tested: 

exactly as on video above

### How Has This Change Been Documented:
-- this is backend only for now