fix: system env variables impacting build caching (#1466)
Currently we include the resolved environment variables in the build
cache key in order to invalidate builds when the user changes something
in their environment which can impact the build.
However, this causes issues with e.g. the `VSCODE_PID` variable which
changes each time vscode is fully closed and restarted, as well as,
presumably, similarly "temporary" variables set by the user's
environment.
This PR changes that behaviour by only including the env variables
configured in the project's `.env` files and Launch Configurations in
the `env` record used throughout the code, including the Build Cache.

### How Has This Been Tested: 
- open an app in Radon
- add some env variable to the launch config
- add a `.env` file with some env variable set
- verify that changing either causes the app to be built again when
restarting