Skip installing iOS app when it is already there (#1464)
This PR implements a fast-path for the case when an app is already
installed on iOS simulator.

App installation is one of the necessary step when starting Radon IDE.
After looking into it, the test-apps typically install within 5-8
seconds on my Mac M2. This can be improved by skipping installation when
the app is already installed.

This PR introduces a mechanism that allows us to determine if app in the
exact installed form already exists:
1) we calculated md5 hash for the build directory that gives us the hash
of the output
2) we use app container location (place in local filesystem where app is
"installed")
3) we place a .dotfile with the hash in that directory
4) when testing whether app is installed, we take the hash of the new
app, and compare it with what's in the .dotfile
5) this works, because when over-installing, simctl wipes out the whole
app container directory and the newly installed app is located in a new
directory (this holds true even if we install exactly the same app
binary)

As a result, we replace 5-8sec installation step (on Mac M2) with the
check which takes 0.3sec (on the same machine)

### Other changes

As part of this PR, we are moving buildHash from `BuildCacheInfo` into
`BuildResult`. The reason is that we can save on calculating the hashes.
We already calculate hash at least once for a produced build anyways.
With the new skipping approach, we'd also need to calculate and verify
it again. This saves us a cost of doing that which can be significant as
the app hashing iterates over files from the .app folder.

### How Has This Been Tested: 
1) install test app on iOS sim
2) restart IDE, see the installation takes less time (we print "skipping
installation" in the logs)
3) test force re-installation which should print "App installed" message
in logs
4) change something in app.json which result in the app rebuilding and
you should see "app installed" message again.