Reactor build to make builds depend on device and also fix cache to rely on device config (#1461)
This PR refactors the build logic to let the build configuration access
the device instance. In particular, `xcodebuild` may need to know the
device it builds for. But also to address two main flows of the build
strategy:
1) The current implementations leaves orphaned entries in the global
state. The reason is that the cache key depends on some parameters that
can be adjusted by the user and in particular we also changed the key
prefix recently when we added support for iPads.
2) We currently use too few parameters for fingerprinting and hence it
may happen that we will attempt to reuse a build that should not be
reused. In particular, we only use a subset of launch config options
while all of them may impact the output of the build process.

To address the above, we are introducing a new concept of build
fingerprint. Build fingerprint is an extension of the fingerprinting
mechanism we used before, but also include the additional parameters
that can impact the build process and are specific to Radon. The idea is
that build fingerpring should represent all input that took part in
producing the build artifact. This includes the project fingerprint
(representing all the native code and dependencies), but also the build
settings and environment constants that are set in the launch
configuration.

To calculate the build fingerprint we therefore take the project
fingerprint and add a deterministic hash of the build configuration. We
also place a platform prefix in front for the fingerprint to be better
readable.

In order to manage the build cache, we are introducing a new GC
mechanism. GC scans the entries in global store and tries to delete ones
that no longer has corresponding backing build artifacts. We also
consider the case where many keys point to a single build artifact, in
which case we cleanup the ones that expect a different artifact hash.

Finally, In order for build config to remain stable, we had to remove
`forceCleanBuild` from it and move it to `BuildOption`. This is an
option that can be set for some builds and not for others, and does not
impact the build artifact but only forces executing the build even if
the fingerprint matches.

### Test plan
1) run iOS and Android builds
2) try force clean build option
3) make sure cached builds are used after restarting the IDE
4) run unit tests