Use destination specifier when building ios (#1471)
This PR changes the set of specifiers we use when building xcode.
Apparently, some of our users reported that under specific conditions
the builds are failing. The root cause turned out to be the fact that
xcodebuild doesn't allow for `ONLY_ACTIVE_ARCH` to ever be overwritten
manually. This flag is important for certain xcode projects when
building for simulator in case the projects contain targets that cannot
build for simulator (i.e. apple watch targets). When specifying
simulator as destination, this is detected automatically by xcode build
and the `ONLY_ACTIVE_ARCH` is set. Apparently this flag is AFAIK the
only one that cannot be overwritten by the options passed to xcodebuild
(https://developer.apple.com/documentation/xcode/build-settings-reference#Build-Active-Architecture-Only)
– it can only be switched off with destination specifier.

As a consequence, we need to use `destination` for xcode build, but this
wasn't as easy. In Radon we use a custom device set. While simctl
supports this via `--set` flag, `xcodebuild` always uses the default
device set. To workaround that, we need to use a device that exists in
the default set. Before starting the build we proceed as follows:
1) we take the runtime id as argument for the build and look for
existing device with this runtime id – in case of iOS, we can pick any
device, as we override the target family later on and thus we don't need
to handle iPhones and iPads separately
2) in no such device exists (which is unlikely), we create a temporary
device that we delete after the build process is over.

### How Has This Been Tested: 
1) run iOS build, see in the logs that it uses the destination specifier
2) this has been verified on the users' project where the problem with
ONLY_ACTIVE_ARCH surfaced