refactor: move to legend state: Multimedia, zoom controls and initialized  (#1474)
This PR continues the move in the direction of the new state management
solution, by moving:
- multimedia state - we create a new `Multimedia` class that is
responsible for capturing screenshots screen recordings and replays.
This new class instance leaves in the `deviceSession` which allows as to
record multiple devices at once
- initialized - a simple state informing frontend about the end of
project initialization
- zoom controls 

### How Has This Been Tested: 

- run `reactNative81` bare application and test the multimedia features
- test zoom controls

### How Has This Change Been Documented:

internal change altho because of multidevice recording the users might
feel the difference.