feat: Improve GitHub Actions workflow for running tests (#1477)
This PR expands the GitHub Actions workflow with new options for running
tests. It adds an option to enable or disable test execution recording
and allows selecting which specific tests to run. For tests executed
locally on a VM, there is also an option to select which tests to
execute.
The PR also makes the GitHub Actions runner setup faster.

### How Has This Been Tested
- GitHub Actions workflows were run with different combinations of
options (recording enabled/disabled, different sets of tests) to verify
correct behavior.
- Local VM test executions were performed with selected tests.  
- To reproduce:  
1. Run the GitHub Actions workflow and select the desired test options
on branch `@KeyJayY/ImproveGHActions`.
2. Execute tests locally on a VM using the provided script with the
appropriate arguments.

Command for running on a local VM:  
`npm run run-tests-on-VM -- <app-directory-name> <test-numbers> `

- `<app-directory-name>`: name of the directory containing the app from
this repository:
[radon-ide-test-apps](https://github.com/software-mansion-labs/radon-ide-test-apps)
- `<test-numbers>`: numbers assigned to test suites.

Currently available test suites:  
  1 - 01-smoke.test.js  
  2 - 02-GUIButtons.test.js  
  3 - 03-manageDevices.test.js  

Examples:  
- `npm run run-tests-on-VM -- react-native-78 2 1` - runs GUIButtons and
Smoke tests (in this order) on the app `react-native-78`.
- `npm run run-tests-on-VM -- react-native-78 --exclude 2` excludes test number 2.
- `npm run run-tests-on-VM -- react-native-78` - runs all available
tests on the app `react-native-78`.
- `npm run run-tests-on-VM` - runs all tests on the default test app.


### How Has This Change Been Documented
No documentation yet.