feat: send files UI (#1480)
This PR is a continuation of #1453 adding a UI whenever a user dragges
files over the simulator preview.
It adapts the UI changes made in #1456 but moves file transfer state
management to legend state.

The UI mimics in style the debug overlay interface.

🚨 Important! You need to hold shift to drag files onto the panel
VSCode displays a "Hold shift to drop into editor" message when Preview
is in the Editor Tab
but it doesn't show this when Preview is in the Side Panel


https://github.com/user-attachments/assets/7963a836-72de-449c-ab6e-c43cd7d58c32

### How Has This Change Been Documented:
WIP

---------

Co-authored-by: kacperkapusciak <kacper.kapusciak@swmansion.com>