feat: Add network panel and screenshot tests, refactor code (#1509)
This pull request introduces the first tests for the network panel, as
well as tests for saving screenshots and screen recordings. It also
refactors and better organizes the code to make future maintenance
easier.

### How Has This Been Tested
- Verified on GitHub Actions CI  
- Verified locally by running the test suite  

To reproduce:

**Locally:**
- Run the command `npm run prepare-and-run-tests` to run the entire test
suite, or
`npm run prepare-and-run-tests -- react-native-81 5 6` to only run
network panel and screenshots tests.
- Tests can also be run on a local VM using the command `npm run
run-tests-on-VM`, after setting up the VM.

**Remotely:**
- Tests can be run through GitHub Actions on the branch
`@KeyJayY/NetworkTesting`.

### How Has This Change Been Documented
- Not documented yet.