Don't crash when ios device shutdown command fails (#1519)
Because of some timing issues with handling devices, the IDE may
sometimes land in the situation when it thinks the device is running
while it's been shut down already. A normal way of trying to recover
from such situation was to use the reload option. However, since reload
triggers a call to the shutdown command from `simctl` and that command
subsequently fails with non-0 error code when the device isn't running,
we'd end up stuck in the reload process.

This PR adds options to shutdown command that prevents it from throwing
an error in case of failure.

We also restructure the exec subprocess implementation such that it
still prints the errors from failed processes even if they don't throw.

### How Has This Been Tested: 
1. Run app
2. Use simctl from terminal to shutodown the device
3. Use reload to try to recover
4. Before you'd get stuck on the reload, now it sucessfully restarts the
device and the app is launched.