refactor: split Devtools into Server and Connection (#1497)
Splits the current `Devtools` class into `DevtoolsServer` (a server the
RN application connects to) and `DevtoolsConnection` (an active
connection between a running app and Radon, created when the app
connects to the `DevtoolsServer`).

This is motivated by two things:
- we may wish to reuse the same `DevtoolsServer` between different
`Application`/`DeviceSession`s. We're planning on reusing the same Metro
server for all currently running applications, but because we use Metro
to provide the port the application should connect to React DevTools
with, we need to run a single Server for all apps using that Metro
instance.
- we may want to move the React DevTools communication from the
WebSocket implementation we currently use to CDP, which is the default
for the new RN DevTools (RN >0.75). This abstracts the communication
channel away from how Radon uses the RDT to talk with the app.

### How Has This Been Tested: 
- open expo-52-prebuild-with-plugins test app in Radon
- check that the devtools (Network, Redux, React Query) still work
- check that element inspection still works
- open an app running Expo Go on Android
- check that debugger works correctly after JS Reloads

### How Has This Change Been Documented:
Internal change